EMSUSD-808 - Updated the code based on the feedback
Also made ufeValueToVtValue and vtValueToUfeValue available from UFE V3
