+++ lib/usdUfe/ufe/UsdUndoClearSceneItemMetadataCommand.h
@@ -20,32 +20,23 @@
 #include <usdUfe/ufe/UsdSceneItem.h>
 #include <usdUfe/ufe/UsdUndoableCommand.h>
 
-#include <pxr/usd/usd/prim.h>
-
 #include <ufe/sceneItem.h>
 #include <ufe/undoableCommand.h>
 #include <ufe/value.h>
-//#include <ufe/ufeFwd.h>
 
 namespace USDUFE_NS_DEF {
 
-//! \brief ARAM TODO: undoable commands.
+//! \brief Undoable command to clear the meta data that is set on a scene item
 //
 
-class USDUFE_PUBLIC SetSceneItemMetadataCommand
+class USDUFE_PUBLIC ClearSceneItemMetadataCommand
     : public UsdUfe::UsdUndoableCommand<Ufe::UndoableCommand>
 {
 public:
-    SetSceneItemMetadataCommand(
+    ClearSceneItemMetadataCommand(
         const PXR_NS::UsdPrim& prim,
         const std::string&     group,
-        const std::string&     key,
-        const Ufe::Value&      value);
-
-    SetSceneItemMetadataCommand(
-        const PXR_NS::UsdPrim& prim,
-        const std::string&     key,
-        const Ufe::Value&      value);
+        const std::string&     key = "");
 
     void executeImplementation() override;
 
@@ -54,7 +45,6 @@ private:
     const PXR_NS::SdfPath         _primPath;
     const PXR_NS::TfToken         _group;
     const std::string             _key;
-    const Ufe::Value              _value;
 };
 
 } // namespace USDUFE_NS_DEF
