+++ lib/usdUfe/ufe/Utils.h
@@ -27,11 +27,11 @@
 #include <ufe/types.h>
 #include <ufe/ufe.h>
 
-#ifdef UFE_SCENEITEM_HAS_METADATA
+#ifdef UFE_V3_FEATURES_AVAILABLE
 #include <pxr/base/vt/value.h>
 
 #include <ufe/value.h>
-#endif
+#endif // UFE_V3_FEATURES_AVAILABLE
 
 #include <string>
 
@@ -210,13 +210,13 @@ Ufe::Selection removeDescendants(const Ufe::Selection& src, const Ufe::Path& fil
 USDUFE_PUBLIC
 Ufe::Selection recreateDescendants(const Ufe::Selection& src, const Ufe::Path& filterPath);
 
-#ifdef UFE_SCENEITEM_HAS_METADATA
+#ifdef UFE_V3_FEATURES_AVAILABLE
 //! Converts a UFE Value to a VtValue
 USDUFE_PUBLIC PXR_NS::VtValue ufeValueToVtValue(const Ufe::Value& ufeValue);
 
 //! Converts a VtValue to a UFE Value
 USDUFE_PUBLIC Ufe::Value vtValueToUfeValue(const PXR_NS::VtValue& vtValue);
-#endif
+#endif // UFE_V3_FEATURES_AVAILABLE
 
 //------------------------------------------------------------------------------
 // Verify edit restrictions.
